Update LFCalc.js
Instead of summing a list of column values it is concatenating them. It seems to be treating the numbers as strings.

This approach ensures that even if the raw value is fetched as a string (which can happen with form inputs), it will be correctly converted to a number before summing.